+++ app/pages/competition/club/club.page.ts
@@ -1,19 +1,15 @@
-import {Component, OnInit } from '@angular/core';
-import {Logger} from "../../../providers/logger";
-//import {SearchList, ISearchEvent} from "../../controls/searchList/searchList";
-import {AppRoutingService} from "../../../context/router.context";
-import {CompetitionService} from "../../../providers/leagues/competitions";
-import {ClubService} from "../../../providers/leagues/club";
-import {GradeService} from "../../../providers/leagues/grade";
-import {RegionCache, CompetitionCache, GradeCache, ClubCache} from "../../../providers/leagues/cache";
-import {CompetitionNav} from "../../nav/competition.nav";
-import {CompetitorResult} from "../../templates/competitor.results";
-
 import * as Models from "../../../models/models";
 
-// import 'rxjs/add/operator/map';
-// import 'rxjs/add/operator/zip';
-// import 'rxjs/add/operator/from';
+import {Component, OnInit} from '@angular/core';
+
+import {AppRoutingService} from "../../../context/router.context";
+import {ClubService} from "../../../providers/leagues/club";
+import {CompetitionCache} from '../../../providers/leagues/competitionCache';
+import {CompetitionNav} from "../../nav/competition.nav";
+import {CompetitionService} from "../../../providers/leagues/competitions";
+import {CompetitorResult} from "../../templates/competitor.results";
+import {GradeService} from "../../../providers/leagues/grade";
+import {Logger} from "../../../providers/logger";
 
 @Component({
     selector: "club-list-page",
