+++ app/pages/competition/clubList/club.list.page.ts
@@ -1,17 +1,14 @@
-import {Component, OnInit } from '@angular/core';
-import {Logger} from "../../../providers/logger";
-//import {SearchList, ISearchEvent} from "../../controls/searchList/searchList";
-import {AppRoutingService} from "../../../context/router.context";
-import {CompetitionService} from "../../../providers/leagues/competitions";
-import {ClubService} from "../../../providers/leagues/club";
-import {GradeService} from "../../../providers/leagues/grade";
-import {RegionCache, CompetitionCache, GradeCache, ClubCache} from "../../../providers/leagues/cache";
-import {IClub} from "../../../models/models";
-import {CompetitionNav} from "../../nav/competition.nav";
+import {Component, OnInit} from '@angular/core';
 
-// import 'rxjs/add/operator/map';
-// import 'rxjs/add/operator/zip';
-// import 'rxjs/add/operator/from';
+import {AppRoutingService} from "../../../context/router.context";
+import {ClubService} from "../../../providers/leagues/club";
+import {CompetitionCache} from '../../../providers/leagues/competitionCache';
+import {CompetitionNav} from "../../nav/competition.nav";
+import {CompetitionService} from "../../../providers/leagues/competitions";
+import {GradeService} from "../../../providers/leagues/grade";
+import {IClub} from "../../../models/models";
+import {Logger} from "../../../providers/logger";
+
 @Component({
     selector: "club-list-page",
     //templateUrl: "pages/competition/clubList/page.html",
