+++ app/pages/competition/competition.page.ts
@@ -1,16 +1,15 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
-
-import {Logger} from "../../providers/logger";
-import {CompetitionService} from "../../providers/leagues/competitions";
-
-import {GradeService} from "../../providers/leagues/grade";
-import {RegionCache, CompetitionCache, GradeCache, ClubCache} from "../../providers/leagues/cache";
-
-import {ClubService} from "../../providers/leagues/club";
-import {ICompetition} from "../../models/models"
-import {CompetitionNav} from "../nav/competition.nav";
-import {AppRoutingService} from "../../context/router.context";
 import * as Rx from 'rxjs/Rx';
+
+import { Component, OnDestroy, OnInit } from '@angular/core';
+
+import {AppRoutingService} from "../../context/router.context";
+import {ClubService} from "../../providers/leagues/club";
+import {CompetitionCache} from '../../providers/leagues/competitionCache';
+import {CompetitionNav} from "../nav/competition.nav";
+import {CompetitionService} from "../../providers/leagues/competitions";
+import {GradeService} from "../../providers/leagues/grade";
+import {ICompetition} from "../../models/models"
+import {Logger} from "../../providers/logger";
 
 @Component({
     selector: "Competiton",
