+++ app/pages/competition/grade/competitors/grade.competitors.page.ts
@@ -1,18 +1,19 @@
-import {Component, OnInit} from '@angular/core';
-import {Logger} from "../../../../providers/logger";
-//import {SearchList, ISearchEvent} from "../../controls/searchList/searchList";
-import {AppRoutingService} from "../../../../context/router.context";
-import {CompetitionService} from "../../../../providers/leagues/competitions";
-import {ClubService} from "../../../../providers/leagues/club";
-import {GradeService} from "../../../../providers/leagues/grade";
-import {CompetitorService} from "../../../../providers/leagues/competitors";
-import {RegionCache, CompetitionCache, GradeCache, ClubCache} from "../../../../providers/leagues/cache";
-import * as Models from "../../../../models/models";
-import {CompetitionNav} from "../../../nav/competition.nav";
-import * as Rx from "rxjs";
 import 'rxjs/add/operator/max';
 import 'rxjs/add/operator/distinct';
 
+import * as Models from "../../../../models/models";
+import * as Rx from "rxjs";
+
+import {Component, OnInit} from '@angular/core';
+
+import {AppRoutingService} from "../../../../context/router.context";
+import {ClubService} from "../../../../providers/leagues/club";
+import {CompetitionCache} from '../../../../providers/leagues/competitionCache';
+import {CompetitionNav} from "../../../nav/competition.nav";
+import {CompetitionService} from "../../../../providers/leagues/competitions";
+import {CompetitorService} from "../../../../providers/leagues/competitors";
+import {GradeService} from "../../../../providers/leagues/grade";
+import {Logger} from "../../../../providers/logger";
 
 @Component({
     selector: "grade-competitors-page",
