+++ app/pages/competition/gradeList/grade.list.page.ts
@@ -1,14 +1,15 @@
-import {Component, OnInit, OnDestroy} from '@angular/core';
-import {Logger} from "../../../providers/logger";
-//import {SearchList, ISearchEvent} from "../../controls/searchList/searchList";
-import {AppRoutingService} from "../../../context/router.context";
-import {CompetitionService} from "../../../providers/leagues/competitions";
-import {ClubService} from "../../../providers/leagues/club";
-import {GradeService} from "../../../providers/leagues/grade";
-import {RegionCache, CompetitionCache, GradeCache, ClubCache} from "../../../providers/leagues/cache";
-import {IGrade} from "../../../models/models";
-import {CompetitionNav} from "../../nav/competition.nav";
 import * as Rx from "rxjs";
+
+import {Component, OnDestroy, OnInit} from '@angular/core';
+
+import {AppRoutingService} from "../../../context/router.context";
+import {ClubService} from "../../../providers/leagues/club";
+import {CompetitionCache} from '../../../providers/leagues/competitionCache';
+import {CompetitionNav} from "../../nav/competition.nav";
+import {CompetitionService} from "../../../providers/leagues/competitions";
+import {GradeService} from "../../../providers/leagues/grade";
+import {IGrade} from "../../../models/models";
+import {Logger} from "../../../providers/logger";
 
 @Component({
     selector: "grade-list-page",
