+++ app/pages/competition/startList/start.list.page.ts
@@ -1,18 +1,17 @@
-import {Component, OnInit} from '@angular/core';
-//import {Router} from "@angular/router-deprecated";
-import {Logger} from "../../../providers/logger";
-import {AppRoutingService} from "../../../context/router.context";
-//import {SearchList, ISearchEvent} from "../../controls/searchList/searchList";
-import {CompetitionService} from "../../../providers/leagues/competitions";
-import {ClubService} from "../../../providers/leagues/club";
-import {GradeService} from "../../../providers/leagues/grade";
-import {RegionCache, CompetitionCache, GradeCache, ClubCache} from "../../../providers/leagues/cache";
-import {Observable, Subscription, Subject} from 'rxjs/Rx';
-import {GroupedObservable} from "rxjs/operator/groupBy";
-
-import {StartListItems} from "./start.list.items.control";
-import {CompetitionNav} from "../../nav/competition.nav";
 import * as Models from "../../../models/models";
+
+import {Component, OnInit} from '@angular/core';
+import {Observable, Subject, Subscription} from 'rxjs/Rx';
+
+import {AppRoutingService} from "../../../context/router.context";
+import {ClubService} from "../../../providers/leagues/club";
+import {CompetitionCache} from '../../../providers/leagues/competitionCache';
+import {CompetitionNav} from "../../nav/competition.nav";
+import {CompetitionService} from "../../../providers/leagues/competitions";
+import {GradeService} from "../../../providers/leagues/grade";
+import {GroupedObservable} from "rxjs/operator/groupBy";
+import {Logger} from "../../../providers/logger";
+import {StartListItems} from "./start.list.items.control";
 
 @Component({
     selector: "start-list-page",
