+++ app/pages/pages.ref.ts
@@ -6,36 +6,20 @@ export { StartListSearchControl } from "./start/start.list.search.control";
 export { StartListControl} from "./start/stat.list.start.control";
 
 export { RegionsPage } from "./regions/regions.page";
-export { RegionPage } from "./region/region.page";
-//level 2
-export { CompetitionPage } from "./competition/competition.page";
+    export { RegionPage } from "./region/region.page";
 
-//level 3
-export { ClubListPage } from "./competition/clubList/club.list.page";
-export { FindCompetitorPage } from "./competition/findCompetitor/find.competitor.page";
-export { GradeListPage } from "./competition/gradeList/grade.list.page";
-export { InformationPage } from "./competition/information/information.page";
-export { StartListPage } from "./competition/startList/start.list.page";
+
+export { CompetitionPage } from "./competition/competition.page";
+    export { ClubListPage } from "./competition/clubList/club.list.page";
+        export {ClubPage} from "./competition/club/club.page";
+    export { FindCompetitorPage } from "./competition/findCompetitor/find.competitor.page";
+    export { GradeListPage } from "./competition/gradeList/grade.list.page";
+        export {GradeCompetitorsPage} from "./competition/grade/grade.page";
+    export { InformationPage } from "./competition/information/information.page";
+    export { StartListPage } from "./competition/startList/start.list.page";
+        export {StartListGradePage } from "./competition/startList/competitors/competitors.page";
 
 //level 4
-export {StartListGradePage } from "./competition/startList/competitors/competitors.page";
-export {ClubPage} from "./competition/club/club.page";
-export {GradeCompetitorsPage} from "./competition/grade/grade.page";
 
 
-// exports = {
-//     StartPage,
-//     RegionsPage,
-//     RegionPage,
 
-//     CompetitionPage,
-//     InformationPage,
-//     StartListPage,
-//     ClubListPage,
-//     GradeListPage,
-//     FindCompetitorPage,
-    
-//     StartListGradePage,
-//     ClubPage,
-//     GradeCompetitorsPage
-// }
