+++ app/providers/providers.ref.ts
@@ -1,0 +1,11 @@
+export {AppRoutingService} from '../context/router.context';
+export { GradeCache } from './leagues/gradeCache';
+export { ClubService } from './leagues/club';
+export { CompetitionCache} from './leagues/competitionCache';
+export { RegionCache } from './leagues/regionCache';
+export { ClubCache } from './leagues/clubCache';
+export { DisplayService } from './leagues/display';
+export { GradeService } from './leagues/grade';
+export { RegionService } from './leagues/leagues';
+export { Logger } from './logger';
+
